apiserver/{client,charms}: do not import types from the client api
Various apiserver packages import types from api packages, _its client_.

- Move ClientInfo to apiserver/params where it belongs.
- Replace the previous apiserver/params.CharmInfo type with a local
  struct; it only needs to have field called `CharmURL`, the rest is
  unimportant.
